rpi/requirements.txt
paho-mqtt>=1.6.1
fastapi>=0.104.0
uvicorn>=0.24.0
plotly>=5.17.0
kaleido>=0.2.1
jinja2>=3.1.2
python-telegram-bot>=20.0
matplotlib>=3.7.0
scipy>=1.11.0
inotify
python-dotenv>=1.0.0
requests>=2.31.0
